use a singular websocket per adapter
