+++ README.md
@@ -42,3 +42,5 @@ Edge, Opera, etc. also use the Chrome Web Store
 - Run `pnpm i`
 - Run `pnpm run build`
 - The output will be in /dist
+
+TODO: prevent generic script from opening a websocket until it finds metadata
