+++ src/extension/content/sites/Bandcamp.ts
@@ -1,6 +1,6 @@
-import { capitalize, timeInSecondsToString } from '../../utils/misc'
-import { querySelector, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { capitalize, timeInSecondsToString } from '../../../utils/misc'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 const site: Site = {
   ready: () => querySelector<boolean, HTMLAudioElement>('audio', (el) => !!el.src, false),
