+++ src/extension/content/sites/Deezer.ts
@@ -1,6 +1,6 @@
-import { getMediaSessionCover } from '../../utils/misc'
-import { querySelector, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { getMediaSessionCover } from '../../../utils/misc'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 // Not reporting duration, position and rating as it seems they disappear once deezer annoys you with its ads
