+++ src/extension/content/sites/Pandora.ts
@@ -1,6 +1,6 @@
-import { capitalize } from '../../utils/misc'
-import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { capitalize } from '../../../utils/misc'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 const site: Site = {
   ready: () =>
