+++ src/extension/content/sites/Spotify.ts
@@ -1,6 +1,6 @@
-import { getMediaSessionCover } from '../../utils/misc'
-import { querySelector, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { getMediaSessionCover } from '../../../utils/misc'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 const site: Site = {
   ready: () => navigator.mediaSession.metadata !== null && querySelector<boolean, HTMLElement>('(.player-controls__buttons button svg path)[3]', (el) => true, false),
