+++ src/extension/content/sites/Youtube.ts
@@ -1,6 +1,6 @@
-import { timeInSecondsToString } from '../../utils/misc'
-import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { timeInSecondsToString } from '../../../utils/misc'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 // TODO: support the tiny player it sometimes gives you in the bottom right corner
