+++ src/extension/content/sites/YoutubeMusic.ts
@@ -1,5 +1,5 @@
-import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../../utils/selectors'
-import { RepeatMode, Site, StateMode } from '../content'
+import { RepeatMode, Site, StateMode } from '../../types'
+import { querySelector, querySelectorEvent, querySelectorEventReport, querySelectorReport } from '../selectors'
 
 let currentCoverUrl = ''
 let lastCoverVideoId = ''
