+++ src/extension/sw/sw.ts
@@ -1,0 +1,5 @@
+import { MessageHandler } from './messaging'
+import { initPort } from './port'
+
+chrome.runtime.onMessage.addListener(MessageHandler)
+initPort()
